snake_oil_no_data: jobs execs changed to symlinks.
